requirements.txt
requests>=2.31.0
paho-mqtt>=1.6.1
typing>=3.7.4.3
anthropic>=0.3.0
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
gunicorn>=20.1.0
click>=8.0.0
pyyaml
opencv-python
pillow